backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.13
botocore==1.37.13
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.4
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
